requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
gunicorn==20.1.0
httpx==0.22.0
pdfplumber==0.5.28
python-docx==0.8.11
pytz==2021.3
python-multipart
